Crate com_core

Build Rust bindings for COM interfaces.

Examples

use ::winapi::HRESULT;
use ::hr::HResult;

use ::com_core::ComResult;
use ::com_core::com::Com;

// Define the COM FFI interface struct and its vtbl struct.

use ::com_core::unknown::{IUnknown, IUnknownVtbl};
com_interface! {
    /// Doc comment.
    pub struct IInterface(IInterfaceVtbl) : IUnknown(IUnknownVtbl);
    {0x582cdb5a-0xe8b9-0x47bb-0xbff8-0x5a18da55bc99}
    pub Foo: unsafe extern "stdcall" fn(
        This: *mut IInterface,
    ) -> HRESULT,
}

// Define and implement the Rust binding for this interface.

use ::com_core::unknown::IUnknownPtr;
com_ptr! {
    /// Doc comment.
    #[derive(Clone, Debug)]
    pub struct IInterfacePtr(IInterface) : IUnknownPtr;
}

impl IInterfacePtr {
    pub fn foo(&self) -> ComResult<()> {
        unsafe {
            let hr: HResult = com_call!(Foo(self)).into();
            hr.result(())
        }
    }
}

// An externally defined com class which implements `IInterface`.

com_class!(_extern Interface, {0xf08b1ee5-0xc953-0x4faf-0x813c-0x9bd256d6fe48}, IInterface);

// Initialize COM and create an com pointer instance for this class.

let com = Com::initialize().unwrap();
let interface_ptr = com.create_instance::<IInterfacePtr, Interface>();

// The above will fail because there's really no such class, but you get the point.
assert!(interface_ptr.is_err());

Reexports

pub extern crate hresult as hr;

Modules

com
COM initialization.

macros
Automate COM plumbing.

unknown
Implements IUnknown.

Macros

com_call!
Calls a virtual function from a COM interface.

com_class!
Implements ComClass given a class, the interface it implements and its class id.

com_interface!
Defines an interface struct and its vtbl struct.

com_ptr!
Defines the rustic wrapper for a COM interface.

com_vtbl!
Defines a vtbl struct.

guid!
Constructs a GUID.

static_guid!
Defines a named static GUID.

Traits

AsComPtr
Access the com interface pointer.

ComCast
Casts an instance to a related type.

ComClass
COM Class structs implement ComClass.

ComClassInterface
Declares the ComInterfaces which are implemented by a ComClass.

ComInherit
Declares a com interface inherits another interface.

ComInterface
COM Interface structs implement ComInterface.

ComPtr
Com pointers wrap the raw com interface pointers and give them a safe rustic interface.

ComVtbl
COM Vtbl structs implement ComVtbl which is little more than a marker.

Type Definitions

ComResult