Struct com_core::com::Com

pub struct Com;

Methods

impl Com

fn initialize() -> ComResult<Com>

Initializes the COM library on the current thread.

See msdn CoInitialize.

fn initialize_ex(flags: CoInit) -> ComResult<Com>

Initializes the COM library for use by the calling thread.

See msdn CoInitializeEx.

fn initialize_security(self) -> ComResult<Com>

Registers security and sets the default security values for the process.

See msdn CoInitializeSecurity.

fn create_instance<P: ComPtr, C: ComClassInterface<P::Interface>>(&self) -> ComResult<P>

Creates a single uninitialized object of the class associated with a specified CLSID.

See msdn CoCreateInstance.

fn set_proxy_blanket<P: ComCast<IUnknownPtr>>(&self, proxy: P) -> ComResult<()>

Sets the authentication information that will be used to make calls on the specified proxy.

See msdn CoSetProxyBlanket.