Module com_core::macros

Automate COM plumbing.

These macros automate the required plumbing to make everything work nicely.

Some work could be done to improve their ergonomics: allow trailing commas, attributes without keywords, ...

GUID macros

Define GUIDs with almost native GUID syntax.

Every number group is required to be prefixed by 0x to indicate they're hexadecimal.

guid!{0xAAAAAAAA-0xBBBB-0xCCCC-0xDDDD-0xEEEEEEEEEEEE}

Creates a GUID instance.

static_guid!(STATIC_GUID, {0xAAAAAAAA-0xBBBB-0xCCCC-0xDDDD-0xEEEEEEEEEEEE});

Expands to a static definition so you avoid having to name its type.

COM FFI macros

These help with plumbing the required FFI structs and associated traits.

com_vtbl!

Defines and implements a vtbl struct.

use ::com_core::unknown::{IUnknownVtbl};

com_vtbl! {
    IInterfaceVtbl: IUnknownVtbl,
    IInterface,
    pub Foo: unsafe extern "stdcall" fn(
        This: *mut IInterface,
    ),
}

com_interface!

Defines and implements an interface struct and its vtbl struct.

use ::com_core::unknown::{IUnknown, IUnknownVtbl};

com_interface! {
    pub struct IInterface(IInterfaceVtbl): IUnknown(IUnknownVtbl);
    {0xAAAAAAAA-0xBBBB-0xCCCC-0xDDDD-0xEEEEEEEEEEEE}
    pub Foo: unsafe extern "stdcall" fn(
        This: *mut IInterface,
    ),
}

com_class!(_extern Class, {0xAAAAAAAA-0xBBBB-0xCCCC-0xDDDD-0xEEEEEEEEEEEE}, IInterface);

Defines a placeholder class (an empty enum), its class id and the interfaces it implements.

COM Rust macros

These help implementing the rustic abstractions for the COM FFI.

com_ptr!

Generates the scaffolding for implementing a rustic interface for an interface.

use ::com_core::unknown::{IUnknown, IUnknownVtbl};

com_interface! {
    pub struct IInterface(IInterfaceVtbl): IUnknown(IUnknownVtbl);
    {0xAAAAAAAA-0xBBBB-0xCCCC-0xDDDD-0xEEEEEEEEEEEE}
    pub Foo: unsafe extern "stdcall" fn(
        This: *mut IInterface,
    ),
}

use ::com_core::unknown::{IUnknownPtr};

com_ptr! {
    #[derive(Clone, Debug)]
    pub struct IInterfacePtr(IInterface): IUnknownPtr;
}

com_call!(Foo(self))

Calls a virtual function on the COM interface.

The first argument should evaulate to an impl AsComPtr.