Trait com_core::ComInterface

pub trait ComInterface: Sized {
    type Vtbl: ComVtbl;
    fn vtbl(&self) -> &Self::Vtbl;
    fn iid() -> &'static IID;
}

COM Interface structs implement ComInterface.

Interfaces have an associated vtbl type and, in COM land, are identified with an IID (I​nterface ID​entifier).

All they contain is a static reference to their vtbl.

Note that it is not legal to 'cast' (transmute, really) the interface to its base interface as this would expose its vtbl with the incorrect type.

Inheritance is declared by ComInherit with an associated type indicating its base interface.

See com_interface! for how to use the provided macros to automate their implementation.

Associated Types

type Vtbl: ComVtbl

Required Methods

fn vtbl(&self) -> &Self::Vtbl

Returns the vtbl.

fn iid() -> &'static IID

Returns the interface id.