Trait com_core::ComPtr

pub trait ComPtr: AsComPtr {
    unsafe fn new(iface: LPVOID) -> Self;
    unsafe fn from_ptr(iface: *mut Self::Interface) -> Self;

    fn cast<P: ComPtr>(&self) -> &P where Self: ComCast<P> { ... }
}

Com pointers wrap the raw com interface pointers and give them a safe rustic interface.

By convention com pointers should be named after their interface name with Ptr added at the end.

Required Methods

unsafe fn new(iface: LPVOID) -> Self

Constructs a new instance from void pointer.

Safety

Must not be null, reinterprets the void pointer as a raw interface pointer.

unsafe fn from_ptr(iface: *mut Self::Interface) -> Self

Constructs a new instance from raw interface pointer.

Safety

Must not be null, should be a valid raw interface pointer.

Provided Methods

fn cast<P: ComPtr>(&self) -> &P where Self: ComCast<P>

Cast the com pointer to a super class.